requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
pickle5
pydantic
PyYAML